test(live): move the executor guard to the executor guard file, and fix a count I broke
Round 3 on #403.

**I regressed a measured number.** Round 1 "corrected" a docstring from 3974 to 3971 --
and 3971 was right *at that moment*, but the same PR then added three tests, which makes
it 3974 again. Measured just now: deleting the `Invalid side` guard gives
`4 failed, 3974 passed, 16 skipped`. The number went DOWN by exactly what the diff added,
which is the tell. In a file whose whole style is measured claims, a stale measurement is
the one defect that matters.

**The new guard was in the wrong file, and that is what forced its shape.**
`test_live_env_base.py` holds contract tests for behaviour shared via `TorchTradeLiveEnv`
-- but these are ORDER EXECUTORS, and `ReplayOrderExecutor` is not a live env at all.
`test_executor_helpers_288.py` is already the executor-surface guard file: it imports the
executor classes at module scope, already parametrizes with `ids=lambda c: c.__name__`,
and its opening docstring names "hedge-mode surface" as one of its three re-fork
precedents. Being in the wrong file is what required `_position_side_free_executors()`,
the function-scoped-import shape round 2 had just deleted from the other file. Moved: 25
added lines become 16, and the wrapper disappears because the imports are already there.
Re-verified after the move -- re-adding `position_side` still fails exactly that venue's
case, by name.

Also: two docstring lines narrating how the parameter was FOUND (commit content, not what
the test pins), and hardcoded `order_executor.py:274`/`:305` line references -- which rot
on the next edit, and `:305` was already wrong: bitget's `hedged=` is at `:324`, `:305`
is a different hedge branch. `(bybit, bitget)` says the same thing and stays true.

The three venue-config imports moved up to join the import block, and
REFORKABLE_FUTURES_CONFIGS down to sit above its only consumer, matching the file's own
convention. Zero net lines; it just stops looking like leftover surgery.

Full suite: 3978 passed, 16 skipped. Diff vs main +41/-42 -- net negative.
